Completed Database Review Assignment
This program has been set to pull weather data from MySQL example_0619
database.  Initially, Orlando and Berlin were not in the system and displayed
the no data message correctly.  I also inserted the data for both cities
into the database so the temperature, pressure, and humidity will be pulled
and displayed to the user.