Completed 5.5 Assignment 7: Object Oriented Programming
Created this program to incorporate inheritance, encapsulation, and
polymorphism.  There is a parent class with 'private' fields, a child class
that uses the same constructor and fields as the parent with successful
calls to the method in the parent class after instantiating one object from
both the super/subclasses respectively.