Finally updated GitHub homepage README
Gave basic overall description of what the repo is and consists of. Added
a screenshot of some newly learned code and how I chose to set it up for
my needs.  Covered the basics of where I'm at with JavaScript and HTML with
link to the file locations for the user to check out