Final Project finished
Full disclosure:  I did not practice version control for this assignment.  I
was very caught up in the code and researching that I never even thought to
start pushing this progress up to this repo, So, I worked until it was
finished offline and this is the only push for this project up to my repo.